[ADD] Move JsonDataAttribute to EmbedIO and document it.
